# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/foglalas": {
            "get": {
                "tags": [
                    "Foglalás"
                ],
                "summary": "Foglalások listázása",
                "description": "Bejelentkezett felhasználó foglalásainak listázása",
                "operationId": "20664e7289efb01235e39cd672c291c6",
                "responses": {
                    "200": {
                        "description": "Sikeres művelet"
                    },
                    "401": {
                        "description": "Nincs bejelentkezve a felhasználó vagy érvénytelen token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "foglalas"
                ],
                "summary": "Foglalás létrehozása",
                "description": "Új foglalás létrehozása a rendszerben.",
                "operationId": "createFoglalas",
                "requestBody": {
                    "description": "Foglalási adatok",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "szolgaltatasnev": {
                                        "type": "string"
                                    },
                                    "letszam": {
                                        "type": "integer"
                                    },
                                    "foglalaskezdete": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "foglalashossza": {
                                        "type": "integer"
                                    },
                                    "megjegyzes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "szolgaltatasnev": "Basic Csomag Plusz",
                                    "letszam": 2,
                                    "foglalaskezdete": "2024-07-12 11:00:00",
                                    "foglalashossza": 4,
                                    "megjegyzes": "Két kamera, mikrofon, fények, live stream."
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sikeres foglalás létrehozása"
                    },
                    "400": {
                        "description": "Érvénytelen foglalási adatok"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/foglalas/{foglalasid}": {
            "delete": {
                "tags": [
                    "foglalas"
                ],
                "summary": "Foglalás törlése",
                "description": "Ez csak bejelentkezett felhasználó által tehető meg.",
                "operationId": "deleteFoglalas",
                "parameters": [
                    {
                        "name": "foglalasid",
                        "in": "path",
                        "description": "Törlendő foglalás azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Érvénytelen foglalás azonosító"
                    },
                    "404": {
                        "description": "Foglalás nem található"
                    },
                    "204": {
                        "description": "Sikeres törlés"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "foglalas"
                ],
                "summary": "Foglalás frissítése",
                "description": "Ez csak bejelentkezett felhasználó által tehető meg.",
                "operationId": "updateFoglalas",
                "parameters": [
                    {
                        "name": "foglalasid",
                        "in": "path",
                        "description": "Frissítendő foglalás azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Frissített foglalás objektum",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "szolgaltatasnev": "Basic Csomag Plusz",
                                    "letszam": 2,
                                    "foglalaskezdete": "2024-06-21 19:00:00",
                                    "foglalashossza": 1,
                                    "megjegyzes": "Két kamera, mikrofon, fények."
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Érvénytelen foglalás adatok"
                    },
                    "404": {
                        "description": "Foglalás nem található"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}